feat: insert activity reservation
