+++ src/controllers/client/activities.ts
@@ -12,7 +12,6 @@ export async function get(req: Request, res: Response) {
 export async function post(req: Request, res: Response) {
   const userId: number = req.user.id;
   const activityId: number = req.body.activityId;
-
   await service.postUserActivity(userId, activityId);
-  res.sendStatus(http.OK);
+  res.sendStatus(http.CREATED);
 }
