+++ src/entities/Activity.ts
@@ -84,31 +84,8 @@ export default class Activity extends BaseEntity {
     return ticketInfos.activities;
   }
 
-  static async postActivity(ticketId: number, activityId: number) {
-    const userActivities = await this.getActivitiesByTicket(ticketId);
-    const newActivityInfos = await this.findOne({ where: { id: activityId } });
-    if (!newActivityInfos) throw InvalidDataError;
-
-    const openVacancies = await this.getOpenVacancies(newActivityInfos.id);
-
-    if (openVacancies === 0) throw InvalidDataError;
-
-    userActivities.forEach((item) => {
-      if (isConflict(item, newActivityInfos)) throw ConflictError;
-    });
-    const ticket = await Ticket.findOne({ where: { id: ticketId } });
+  static async postActivity(ticket: Ticket, newActivityInfos: Activity) {
     newActivityInfos.tickets = [ticket];
-    return newActivityInfos;
+    this.save(newActivityInfos);
   }
-}
-
-function isConflict(item1: Activity, item2: Activity): boolean {
-  if (
-    item1.id === item2.id ||
-    (item1.endTime > item2.startTime &&
-      item1.startTime < item2.endTime &&
-      item1.day.id === item2.day.id)
-  )
-    return true;
-  return false;
 }
